% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altmetric.R
\name{altmetric}
\alias{altmetric}
\title{Gather altmetric metadata from a DOI}
\usage{
altmetric(doi = "10.1016/j.joi.2017.08.007")
}
\arguments{
\item{doi}{is a character. It contains a list of DOIs. A DOI is a persistent identfier of a scholarly document.}
}
\value{
a data frame. Each row contains the full metadata record for each scholarly document.

For more extensive information about Altmetric, please visit: https://www.altmetric.com
}
\description{
It gathers altemtric metadata from a DOI using Altmetric API (www.altmetric.com).
A single research output may live online in multiple websites and can be talked about across dozens of different platforms.
 
Altmetric is a search engine which collects and collates all of this disparate information to obtain an informative view of the
online activity surrounding your scholarly content.
}
\examples{

\dontrun{
doi = "10.1016/j.joi.2017.08.007"

df <- altmetric(doi = doi)
}

}
