% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsApi2df.R
\name{dsApi2df}
\alias{dsApi2df}
\title{Convert data in json dimensions structure to a dataframe}
\usage{
dsApi2df(P)
}
\arguments{
\item{P}{is a list in json dimensions structure downloaded using the function \code{dsApiRequest}.}
}
\value{
a bibliographic dataframe.

To obtain a free access to Dimenions API fro no commercial use, please visit: \href{https://ds.digital-science.com/NoCostAgreement}{https://ds.digital-science.com/NoCostAgreement}

For more extensive information about dimensions API, please visit: \href{https://www.dimensions.ai/dimensions-apis/}{https://www.dimensions.ai/dimensions-apis/}
}
\description{
It converts dimensions data downloaded using DSL API into a data frame
}
\examples{

# token <- dsAuth(username = "my.email@my.domain", password = "mypassword")
# query <- dsQueryBuild(item = "publications", words = "bibliometric*", 
#                        type = "article", categories = "management", 
#                       start_year=1980,end_year = 2020)
# D <- dsApiRequest(token = token, query = query, limit = 50000)
# M <- dsApi2df(D)

}
\seealso{
\code{\link{dsApiRequest}}

\code{\link{dsAuth}}

\code{\link{dsQueryBuild}}
}
