% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsAuth.R
\name{dsAuth}
\alias{dsAuth}
\title{Obtain an API token from dimensions.ai
It generates a token request to dimensions.ai using account and password.}
\usage{
dsAuth(username, password, verbose = FALSE)
}
\arguments{
\item{username}{is a character.}

\item{verbose}{is logical.}

\item{passwrod}{is a character.}
}
\value{
a character cointaining an token o use dimensions API.

For more extensive information about dimensions API, please see: \href{https://www.dimensions.ai/dimensions-apis/}{https://www.dimensions.ai/dimensions-apis/}
}
\description{
Obtain an API token from dimensions.ai
It generates a token request to dimensions.ai using account and password.
}
\examples{

#  token <- dsAuth(username = "my.email@my.domain", password = "mypassword")

}
\seealso{
\code{\link{dsApiRequest}}

\code{\link{dsQueryBuild}}

\code{\link{dsApi2df}}
}
